+++ src/components/radiobutton/index.ts
@@ -1,0 +1,2 @@
+export * from './radiobutton.module';
+export * from './radiobutton.component';